[GUI] Move Font class to gl/font.h from gui/text.h , remove text.h and text.cpp
