+++ src/client/client.h
@@ -5,7 +5,6 @@
 #include "gl/textures.h"
 #include "gl/vertex_array.h"
 #include "gui/gui_master.h"
-#include "gui/text.h"
 #include "maths.h"
 #include "world/chunk_mesh.h"
 #include <SFML/Network/Packet.hpp>
@@ -24,6 +23,7 @@
 #include "renderer/chunk_renderer.h"
 
 class Keyboard;
+class GuiText;
 
 struct BlockUpdate {
     BlockPosition position;
