+++ src/client/gl/font.h
@@ -1,14 +1,13 @@
 #pragma once
+
 #include "../gl/textures.h"
-#include "../gl/vertex_array.h"
 #include <SFML/Graphics/Font.hpp>
-#include <glm/gtc/matrix_transform.hpp>
 
 namespace gl {
-class Shader;
-struct UniformLocation;
-} // namespace gl
-
+/**
+ * @brief Opengl supporting font
+ *
+ */
 class Font {
   public:
     void init(const std::string& fontFile, unsigned bitmapScale);
@@ -31,23 +30,4 @@ class Font {
                                   "IJKLMNOPQRSTUVWXYZ.,!?-+/()[]:;%&`*#=\"";
 };
 
-class Text final {
-  public:
-    void setFont(const Font& font);
-    void setText(const std::string& string);
-    void setCharSize(float size);
-    void setPosition(const glm::vec3& position);
-
-    void render(const gl::UniformLocation& location);
-
-  private:
-    void createGeometry();
-
-    std::string m_text;
-    gl::VertexArray m_vao;
-    float m_scale = 0;
-    const Font* m_font = nullptr;
-    bool m_needsUpdate = false;
-
-    glm::vec3 m_position;
-};
+} // namespace gl
